fix(quality): split the backend apply boundary and clear its findings
backend_calls.py carried both the invocation path and the result-gating path in
one 478-line module with an 8-parameter entry point, a 21-cognitive finalizer,
and three helpers whose dataclasses.replace return type defeated the ApplyResult
annotation. Separate the three concerns it actually had:

- backend_calls.py keeps invocation: admission, plan binding, preparation, and
  the backend call itself. The per-invocation identity and host services move
  into one _BackendCallContext, which brings the entry point back inside the
  parameter bound at all 13 call sites.
- backend_apply_results.py owns gating: finalize, sanitize, revalidate, and
  disclose realization authority, each flattened into its own step.
- backend_result_diagnostics.py owns the refusal diagnostics and snapshot
  contract checks both paths need, so neither imports the other.

Cast the three replace()-based helpers to ApplyResult, which also clears the
argument-type finding at the post-apply call. Re-capture the evidence digests,
including the two implementation-surface digests.
