+++ docs/research/formal-semantic-validation/bundles/retest-v8.json
@@ -118,5 +118,5 @@
   "protocol_sha256": "abf94093e344bf495dfb04e8b0c5985c0beaab8ebb17a75e15c8674fa81b1a7c",
   "revision": "9.0.0",
   "snapshot_path": "docs/research/formal-semantic-validation/execution-snapshot-v8.json",
-  "snapshot_sha256": "ace1f0dd511600f67f2fe653ac91721e1c563124ca00c81200d276801fb54351"
+  "snapshot_sha256": "bae8fc592163b52d9c5238f5df8559180165fe012a981aa34d5b66033e7d3acf"
 }
