+++ docs/research/formal-semantic-validation/execution-snapshot-v8.json
@@ -645,7 +645,7 @@
   "source_state": {
     "base_revision": "9b69eb211afd0785809fc9db053f5799dc36eae9",
     "checkout_state": "modified",
-    "implementation_digest": "a3d917c7c810cbe358f89f32d98d6132c818c8abb4354dc482e494c8f5054c50",
+    "implementation_digest": "f7ce0a7ea866291bdfa87f0a5663021d4a3457887d64afb533dfa32061701358",
     "profile": "python-reference-source/v1"
   },
   "versions": {
