+++ docs/research/specification-coverage/analysis-v7.json
@@ -84,5 +84,5 @@
     }
   ],
   "snapshot_id": "raes-standardized-specification-coverage-issue-1204-v7",
-  "snapshot_sha256": "0b35b523a68b8281904f0dc26a28c1a4172ecf4252f64897985ee681a4258d37"
+  "snapshot_sha256": "2f736b0cfdd20fcb6d26e55bd945313c512cc8a7309158bc9c12bc81b2dfb870"
 }
