+++ docs/research/specification-coverage/bundles/raes-standardized-specification-coverage-issue-1204-v7.json
@@ -1,10 +1,10 @@
 {
   "analysis_path": "docs/research/specification-coverage/analysis-v7.json",
-  "analysis_sha256": "5745d9aa0526706e00ca981c32ed649ac21f6afb3af3d1529bd3af8a1167f283",
+  "analysis_sha256": "267d1479c78544ff70077a01190048b93067bd203f56c58c07e43107c6138da3",
   "bundle_id": "raes-standardized-specification-coverage",
   "protocol_path": "docs/research/specification-coverage/protocol-v1.json",
   "protocol_sha256": "e97a19e643e94c9e589dca823a63c6ce49d3329fe2a3cb888ab630838ed93125",
   "revision": "7.0.0",
   "snapshot_path": "docs/research/specification-coverage/execution-snapshot-v7.json",
-  "snapshot_sha256": "72f0ffd81060d106d1fc0e3f78a233e9c4ed11b0277aa0dadfb36d70423a2bcb"
+  "snapshot_sha256": "55b57a6022e46d4c76d3a0111752f14b11b4048a941ecd4606dc483c13e3a7cc"
 }
