+++ docs/research/specification-coverage/execution-snapshot-v7.json
@@ -641,12 +641,12 @@
   "execution_status": "complete",
   "implementation_surfaces": [
     {
-      "content_sha256": "a858b83d10a074e40ffc9babad12f80cfa0cc1483ba52a47bbef9e03fffac3b9",
+      "content_sha256": "7dcb0b0959f75981ea722210570181e7f31a940367278ddd9ca8ea9966694c3c",
       "path": "implementations/python/packages/raes_contracts",
       "surface_id": "contract-models"
     },
     {
-      "content_sha256": "6d8f5c8e54a5a203bb90da569fbf542d8c2fa5fe83210e64b7599b36ac78faa3",
+      "content_sha256": "10e6121b0f51f46cfabb6519fa82c58c40267760bb41af90e01a16d0c988bf7a",
       "path": "implementations/python/packages/raes_processor",
       "surface_id": "processor-pipeline"
     },
@@ -671,7 +671,7 @@
   "source_state": {
     "base_revision": "9b69eb211afd0785809fc9db053f5799dc36eae9",
     "checkout_state": "modified",
-    "implementation_digest": "a3d917c7c810cbe358f89f32d98d6132c818c8abb4354dc482e494c8f5054c50",
+    "implementation_digest": "f7ce0a7ea866291bdfa87f0a5663021d4a3457887d64afb533dfa32061701358",
     "profile": "python-reference-source/v1"
   }
 }
