+++ implementations/python/packages/raes_runtime/control_plane_execution.py
@@ -21,7 +21,7 @@ from raes_contracts.runtime_state import (
     operation_terminal_diagnostics,
 )
 
-from .backend_calls import _call_backend_apply, _call_backend_diagnostics, _RealizationApplyContext
+from .backend_calls import _BackendCallContext, _call_backend_apply, _call_backend_diagnostics, _RealizationApplyContext
 from .backend_observation_calls import _call_backend_apply_with_observation, _ObservationApplyRequest
 from .control_plane_mutation import control_plane_mutation
 from .control_plane_operation_context import operation_admission_context
@@ -52,7 +52,9 @@ def apply_authorized_participant_action(
         address=address,
         snapshot=snapshot,
         realization=participant_effect_authority(request, snapshot),
-        information_state_context_resolver=information_state_context_resolver,
+        call=_BackendCallContext(
+            information_state_context_resolver=information_state_context_resolver,
+        ),
     )
 
 
@@ -139,10 +141,12 @@ def _execute_participant_action_locked(
             address=address,
             snapshot=control_plane._snapshot,
             realization=participant_effect_authority(request, control_plane._snapshot),
-            information_state_context_resolver=getattr(
-                control_plane,
-                "_information_state_context_resolver",
-                None,
+            call=_BackendCallContext(
+                information_state_context_resolver=getattr(
+                    control_plane,
+                    "_information_state_context_resolver",
+                    None,
+                ),
             ),
         )
     final_state = OperationState.SUCCEEDED if result.success else OperationState.FAILED
